fix: Require deterministic signing in the ed25519 WebCrypto implementation. Otherwise, use noble implementation. Fixes identity usage in Safari.
